logfile: Don't check file stat and rollback every time
We want to call logfile for iterative filling to allow logfile
to produce partial results in the future.  To allow that, we
need to allow that there might still be some unprocessed loglines
in the buffer.  Keep the buffer around and only check for file
wonkiness if the previous call left no detritus.